Fix MinGW cross-compile type conflicts
- types.h: Always use int32_t/uint32_t (not long) for int32/uint32/dword
- types.h: Guard boolean typedef with #ifndef _WIN32 (windows.h defines it)
- types.h: Add _TYPES_H define so audio_stub.h can skip duplicate typedefs
- audio_stub.h: Skip type definitions when types.h already included
- GAME.C: Fix _argc type to match util_lib.h declaration (int32, not int)

+++ source/GAME.C
@@ -7376,9 +7376,9 @@ int main(int argc,char **argv)
 
     /* Propagate command-line args to MMULTI.C for network init */
     {
-        extern int _argc;
+        extern int32 _argc;
         extern char **_argv;
-        _argc = argc;
+        _argc = (int32)argc;
         _argv = argv;
     }
